+++ DAL/DBMODEL/PurchaseOrderDetail.cs
@@ -10,14 +10,14 @@ namespace DAL.DBMODEL
         public int Id { get; set; }
         public string ItemName { get; set; }
         public int ProductId { get; set; }
-        public int PurchasePrice { get; set; }
-        public long? SalePrice { get; set; }
+        public decimal PurchasePrice { get; set; }
+        public decimal? SalePrice { get; set; }
         public int Qty { get; set; }
         public int? StockId { get; set; }
         public int PurchaseOrderId { get; set; }
         public DateTime? ExpiryDate { get; set; }
         public DateTime? StartDate { get; set; }
-        public int? Discount { get; set; }
-        public int? Total { get; set; }
+        public decimal? Discount { get; set; }
+        public decimal? Total { get; set; }
     }
 }
